add mc simulator test
